requirements.txt
torchvision
transformers>=4.25.1 
scikit-image
datasets
tqdm
regex
pyyaml
numpy
einops
ftfy
axial_positional_embedding
accelerate
timm
einops_exts
x_transformers==0.28.4
ninja
matplotlib
torchdiffeq
python-etcd
Jinja2
joblib
tensorboard